+++ src/esm/utils/dom/guiSizeToggler.js
@@ -5,7 +5,8 @@
  * @date     2021-12-13
  * @modified 2022-01-10
  * @modified 2024-06-25 Ported to typescript and moved to utils/dom (was located inside demos before).
- * @version  1.1.0
+ * @modified 2024-10-02 Added a transition time.
+ * @version  1.1.1
  */
 /**
  * @param {GUI} gui - The GUI (compatible with dat.gui and lil-gui).
@@ -29,7 +30,7 @@ const applyGuiSize = (gui, guiDoubleSize, cssProps) => {
     }
 };
 export const guiSizeToggler = (gui, config, cssProps) => {
-    gui.domElement.style["transition"] = "scale 2s";
+    gui.domElement.style["transition"] = "transform 0.5s";
     return {
         update: function () {
             applyGuiSize(gui, Boolean(config.guiDoubleSize), cssProps);